+++ M2/Macaulay2/d/engine.dd
@@ -97,7 +97,7 @@ header "
     #include <hash.hpp>
     #include <monomial/monomial.hpp>
     #include <relem.hpp>
-    #include <matrix.hpp>
+    #include <matrices/matrix.hpp>
     #include <basic-mutable-matrices/mat.hpp>
 ";
