+++ M2/Macaulay2/e/Eschreyer.cpp
@@ -1,7 +1,7 @@
 // Copyright 1999  Michael E. Stillman
 
 #include "Eschreyer.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "monoid.hpp"
 #include "text-io.hpp"
 #include "groebner-computations/gbring.hpp"
