+++ M2/Macaulay2/e/Eschreyer.hpp
@@ -5,7 +5,7 @@
 #include "rings/polyring.hpp"
 #include "groebner-computations/gbring.hpp"
 #include "free-modules/schreyer-orders.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "monoid.hpp"
 #include "comp.hpp"
