+++ M2/Macaulay2/e/LLL.cpp
@@ -3,7 +3,7 @@
 #include "LLL.hpp"
 
 #include "relem.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "text-io.hpp"
 #include "interrupted.hpp"
