+++ M2/Macaulay2/e/NAG.cpp
@@ -19,7 +19,7 @@
 #include "interface/NAG.h"
 #include "basic-mutable-matrices/lapack.hpp"
 #include "matrices/matrix-con.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "rings/poly.hpp"
 #include "relem.hpp"
