+++ M2/Macaulay2/e/NCResolutions/nc-res-computation.hpp
@@ -5,7 +5,7 @@
 
 #include "comp-res.hpp"
 #include "NCAlgebras/FreeAlgebraQuotient.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "matrices/matrix-con.hpp"
 
 class NCResComputation : public ResolutionComputation
