+++ M2/Macaulay2/e/bibasis/involutive.hpp
@@ -15,7 +15,7 @@
 #include "pcomparator.hpp"
 #include "qset.hpp"
 #include "tset.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "matrices/matrix-con.hpp"
 
 namespace BIBasis
