+++ M2/Macaulay2/e/bibasis/launcher.hpp
@@ -11,7 +11,7 @@
 #define BIBASIS_LAUNCHER_HPP
 
 #include "involutive.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 
 namespace BIBasis
 {
