+++ M2/Macaulay2/e/cra.cpp
@@ -6,7 +6,7 @@
 #include "error.h"
 #include "matrices/matrix-con.hpp"
 #include "free-modules/freemod.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "monoid.hpp"
 #include "rings/poly.hpp"
 #include "rings/ring.hpp"
