+++ M2/Macaulay2/e/free-modules/freemod.cpp
@@ -4,7 +4,7 @@
 #include "free-modules/freemod.hpp"
 #include "comb.hpp"
 #include "rings/polyring.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "matrices/matrix-con.hpp"
 #include "Eschreyer.hpp"
 #include "groebner-computations/gbring.hpp"
