+++ M2/Macaulay2/e/free-modules/schreyer-orders.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 #include "free-modules/schreyer-orders.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "comb.hpp"
 #include "rings/polyring.hpp"
 #include "Eschreyer.hpp"
