+++ M2/Macaulay2/e/gb-f4/PolynomialList.cpp
@@ -1,5 +1,5 @@
 #include "PolynomialList.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "matrices/matrix-stream.hpp"
 
 namespace newf4 {
