+++ M2/Macaulay2/e/groebner-computations/gb-sugarless.hpp
@@ -3,7 +3,7 @@
 #define _gbinhom_hh_
 
 #include "relem.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "rings/polyring.hpp"
 #include "groebner-computations/comp-gb.hpp"
 #include "groebner-computations/reducedgb.hpp"
