+++ M2/Macaulay2/e/groebner-computations/gb-toric.hpp
@@ -5,7 +5,7 @@
 #define _gbbinom_hh_
 
 #include "groebner-computations/comp-gb.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 
 /////////////////////
 // Data structures //
