+++ M2/Macaulay2/e/hermite.hpp
@@ -3,7 +3,7 @@
 #define _hermite_hh_
 
 #include "relem.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "rings/polyring.hpp"
 #include "groebner-computations/comp-gb.hpp"
 #include "rings/ZZ.hpp"
