+++ M2/Macaulay2/e/interface/matrix.cpp
@@ -16,7 +16,7 @@
 #include "interface/monoid.h"
 #include "basic-mutable-matrices/mat.hpp"
 #include "matrices/matrix-con.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "mutable-matrices/mutablemat-defs.hpp"
 #include "relem.hpp"
 #include "rings/ring.hpp"
