+++ M2/Macaulay2/e/resolutions/res-a2.hpp
@@ -3,7 +3,7 @@
 #define _gb2_hh_
 
 #include "relem.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "rings/polyring.hpp"
 #include "comp-res.hpp"
 #include "hilb.hpp"
