+++ M2/Macaulay2/e/rings/localring.cpp
@@ -10,7 +10,7 @@
 #include "groebner-computations/gbring.hpp"
 #include "relem.hpp"
 #include "debug.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "matrices/matrix-con.hpp"
 #include "mutablecomplex.hpp"
 #include "exceptions.hpp"
