+++ M2/Macaulay2/e/schreyer-resolution/res-f4-m2-interface.cpp
@@ -21,7 +21,7 @@
 #include "interface/groebner.h"                           // for rawMinimalB...
 #include "basic-mutable-matrices/mat.hpp"                                        // for MutableMatrix
 #include "matrices/matrix-con.hpp"                                 // for MatrixConst...
-#include "matrix.hpp"                                     // for Matrix
+#include "matrices/matrix.hpp"                                     // for Matrix
 #include "monoid.hpp"                                     // for Monoid
 #include "newdelete.hpp"                                  // for newarray
 #include "rings/polyring.hpp"                                   // for PolynomialRing
