+++ M2/Macaulay2/e/unit-tests/QuotientRingTest.cpp
@@ -4,7 +4,7 @@
 #include <gtest/gtest.h>
 #include "util-polyring-creation.hpp"
 #include "RingElem.hpp"
-#include "matrix.hpp"
+#include "matrices/matrix.hpp"
 #include "debug.hpp"
 // Step 1: idealFromStrings
 TEST(IdealCreation, fromStrings)
